LPS-80123 (announcements-uad) Add base implementation class for UADExporter
